+++ tasty-dejafu/Test/Tasty/DejaFu.hs
@@ -25,7 +25,7 @@ module Test.Tasty.DejaFu
   -- @instance IsOption Bounds@
   -- @instance IsOption MemType@
 
-  -- * Property testing
+  -- * Unit testing
     testAuto
   , testDejafu
   , testDejafus